requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu116
torch==1.12.1+cu116
torchvision==0.13.1+cu116
torchaudio==0.12.1
numpy==1.22.3
scikit-learn==1.1.2
hydra-core==1.2.0